Environment, Module, and Pipe Browser.

Interrupt the execution of an expression or a pipe and allow the inspection of the environment where browser was called from.

browser(...)

Arguments

...
Further arguments to be passed to browser.

Details

See browser.

Examples

not_run({
  library(magrittr)
  (function() {
    "foobar" %>%
      browser() %>%
      print
  })()
})

reset()

#> [2018-12-02T17:13:56 CET] Resetting modulr state ... OK

define("foo", NULL, function() {foo <- "bar"; browser()})

#> [2018-12-02T17:13:56 CET] Defining 'foo' ... OK

## Not run: make()

reset()

#> [2018-12-02T17:13:56 CET] Resetting modulr state ... OK

define("foobar", NULL, function() {
  library(magrittr)
  "foobar" %>%
    browser %>%
    print
})

#> [2018-12-02T17:13:56 CET] Defining 'foobar' ... OK

## Not run: make()