Plot Dependencies.

Plot the directed acyclic graph (DAG) of modules and dependencies.

plot_dependencies(group, regexp, reserved = TRUE,
  render_engine = sankey_engine, ...)

Arguments

group
A character vector of module names (cf. define) to include as a subset of the graph nodes.

regexp
A regular expression. If not missing, the regular expression is used to filter the names of the modules to be plotted.

reserved
A flag. Should special modules with a reserved name be considered?

render_engine
A function. Rendering engine used to plot the dependencies.

...
Further arguments to be passed to render_engine.

See also

define and reset.

Examples

reset()

#> [2018-12-02T17:14:08 CET] Resetting modulr state ... OK

define("foo", NULL, function() NULL)

#> [2018-12-02T17:14:08 CET] Defining 'foo' ... OK

define("bar", list(m = "modulr"), function(m) NULL)

#> [2018-12-02T17:14:08 CET] Defining 'bar' ... OK

#> Warning: [2018-12-02T17:14:08 CET] Possibly unused dependency in ‘bar’: ‘m’.

define("foobar", list(f = "foo", b = "bar"), function(f, b) NULL)

#> [2018-12-02T17:14:08 CET] Defining 'foobar' ... OK

#> Warning: [2018-12-02T17:14:08 CET] Possibly 2 unused dependencies in ‘foobar’: ‘f’, ‘b’.

define("foobuzz", list(f = "foo", b = "buzz"), function(f, b) NULL)

#> [2018-12-02T17:14:08 CET] Defining 'foobuzz' ... OK

#> Warning: [2018-12-02T17:14:08 CET] Possibly 2 unused dependencies in ‘foobuzz’: ‘f’, ‘b’.

define("fizz", list(f1 = "foobar", f2 = "foobuz"), function(f1, f2) NULL)

#> [2018-12-02T17:14:08 CET] Defining 'fizz' ... OK

#> Warning: [2018-12-02T17:14:08 CET] Possibly 2 unused dependencies in ‘fizz’: ‘f1’, ‘f2’.

wait <- function() invisible(readline(prompt="Press [enter] to continue"))
plot_dependencies("foobar", reserved = FALSE)
## Not run: wait()
plot_dependencies("foobar")
## Not run: wait()
plot_dependencies(reserved = FALSE)
## Not run: wait()
plot_dependencies()