Prepare and Release a Modulr Gear as a Gist on GitHub.

Prepare and release a modulr gear as a gist on GitHub.

release_gear_as_gist(name = .Last.name, load = TRUE, update = TRUE,
  update_first_found = TRUE, per_page = 100L, max_pages = 10L,
  endpoint = "https://api.github.com", browse = TRUE)

Arguments

name
A string (character vector of lenght one).

A module name can contain letters, figures and some special characters, namely _, -, and /. The latter is a namespace separator.

Names containing /mock/, /mocks/, /test/, /tests/, /example/, or /examples/ have a special meaning related to code testing and examples.

The name "modulr" corresponds to a special module and is therefore reserved.

load
A flag. Should an undefined module be implicitely loaded?

update
A flag or a string (character vector of length one). Should an existing gist be updated (flag)? Or gist ID to update (string).

update_first_found
A flag. Should the first corresponding module name be updated?

per_page
A scalar (integer vector of length one). Number of gists per page requested.

max_pages
A scalar (integer vector of length one). Upper bound for the number of pages requested.

endpoint
A string (character vector of length one). GitHub API Endpoint URL.

browse
A flag. Should the created or updated gist be opened in the default browser?

See also

define, prepare_gear, and reset.

Examples

not_run({
  library(gistr)
  reset()
  define("foo", NULL, function() NULL)
  Sys.setenv("GITHUB_PAT" = "your Personal Access Token here")
  gist_auth(reauth = TRUE)
  release_gear_as_gist("foo")
})