With (...) Packages.

Temporarily unload and detach packages, and optionally use a specific packages library for the module.

with_no_packages(code, ignore = NULL)

without_packages(code, ignore = NULL)

with_packages(lib_path, code, ...)

with_module_packages(code, ...)

with_namespace_packages(namespace, code, ...)

Arguments

code
Any object. Code to execute in the temporary environment.

ignore
A vector of character strings. Packages (in addition to base packages) to ignore in the process.

lib_path
A string (character vector of length one) containing the path of the packages library.

...
Further arguments to be passed to with_no_packages.

namespace
A string (character vector of length one) containing the namespace of the module to be used as path for the packages library.

Value

The result of the evaluation of the code argument.

Details

Packages are temporarily unloaded and detached from the search path. When restored, the ordering of the packages on the search path is preserved. Base packages and packages passed to the ignore parameter are ignored in the process. By default, the vector of names of ignored packages is taken from the modulr.ignore_packages R option. with_packages uses the library given by lib_path. with_module_packages uses a dedicated library for the module, using the module name as path. with_namespace_packages uses a shared library for all the modules under the given namespace parameter.

Warning

This is an experimental feature subject to changes.

See also

withr for examples of 'with_' methods, getOption, library, options, and search.

Examples

print(sessionInfo())

#> R version 3.2.5 (2016-04-14)
#> Platform: x86_64-pc-linux-gnu (64-bit)
#> Running under: Red Hat Enterprise Linux
#> 
#> locale:
#>  [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C              
#>  [3] LC_TIME=en_DK.UTF-8        LC_COLLATE=en_US.UTF-8    
#>  [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8   
#>  [7] LC_PAPER=en_US.UTF-8       LC_NAME=C                 
#>  [9] LC_ADDRESS=C               LC_TELEPHONE=C            
#> [11] LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C       
#> 
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base     
#> 
#> other attached packages:
#> [1] memoise_1.1.0     modulr_0.1.7.9180
#> 
#> loaded via a namespace (and not attached):
#>  [1] igraph_1.0.1          Rcpp_0.12.11          rstudioapi_0.6       
#>  [4] whisker_0.3-2         knitr_1.16            xml2_1.1.1           
#>  [7] magrittr_1.5          roxygen2_6.0.1        devtools_1.12.0.9000 
#> [10] usethis_0.0.0.9000    pkgload_0.0.0.9000    R6_2.2.1             
#> [13] chorddiag_0.1.1       rlang_0.1.1.9000      pooh_0.3-2           
#> [16] httr_1.2.1            stringr_1.2.0         networkD3_0.2.13     
#> [19] tools_3.2.5           withr_1.0.2           htmltools_0.3.6      
#> [22] commonmark_1.2        yaml_2.1.14           assertthat_0.2.0.9000
#> [25] rprojroot_1.2         digest_0.6.12         tibble_1.3.3         
#> [28] pkgdown_0.1.0.9000    crayon_1.3.2          callr_1.0.0          
#> [31] purrr_0.2.2.2         htmlwidgets_0.8       testthat_1.0.2       
#> [34] evaluate_0.10         rmarkdown_1.5         stringi_1.1.5        
#> [37] desc_1.1.0            backports_1.1.0       highlight_0.4.7.1    
#> [40] jsonlite_1.5         

not_run({
  with_no_packages({
    message("Look mum, no packages!")
    print(sessionInfo())
  })
})
print(sessionInfo())

#> R version 3.2.5 (2016-04-14)
#> Platform: x86_64-pc-linux-gnu (64-bit)
#> Running under: Red Hat Enterprise Linux
#> 
#> locale:
#>  [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C              
#>  [3] LC_TIME=en_DK.UTF-8        LC_COLLATE=en_US.UTF-8    
#>  [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8   
#>  [7] LC_PAPER=en_US.UTF-8       LC_NAME=C                 
#>  [9] LC_ADDRESS=C               LC_TELEPHONE=C            
#> [11] LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C       
#> 
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base     
#> 
#> other attached packages:
#> [1] memoise_1.1.0     modulr_0.1.7.9180
#> 
#> loaded via a namespace (and not attached):
#>  [1] igraph_1.0.1          Rcpp_0.12.11          rstudioapi_0.6       
#>  [4] whisker_0.3-2         knitr_1.16            xml2_1.1.1           
#>  [7] magrittr_1.5          roxygen2_6.0.1        devtools_1.12.0.9000 
#> [10] usethis_0.0.0.9000    pkgload_0.0.0.9000    R6_2.2.1             
#> [13] chorddiag_0.1.1       rlang_0.1.1.9000      pooh_0.3-2           
#> [16] httr_1.2.1            stringr_1.2.0         networkD3_0.2.13     
#> [19] tools_3.2.5           withr_1.0.2           htmltools_0.3.6      
#> [22] commonmark_1.2        yaml_2.1.14           assertthat_0.2.0.9000
#> [25] rprojroot_1.2         digest_0.6.12         tibble_1.3.3         
#> [28] pkgdown_0.1.0.9000    crayon_1.3.2          callr_1.0.0          
#> [31] purrr_0.2.2.2         htmlwidgets_0.8       testthat_1.0.2       
#> [34] evaluate_0.10         rmarkdown_1.5         stringi_1.1.5        
#> [37] desc_1.1.0            backports_1.1.0       highlight_0.4.7.1    
#> [40] jsonlite_1.5         

print(sessionInfo())

#> R version 3.2.5 (2016-04-14)
#> Platform: x86_64-pc-linux-gnu (64-bit)
#> Running under: Red Hat Enterprise Linux
#> 
#> locale:
#>  [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C              
#>  [3] LC_TIME=en_DK.UTF-8        LC_COLLATE=en_US.UTF-8    
#>  [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8   
#>  [7] LC_PAPER=en_US.UTF-8       LC_NAME=C                 
#>  [9] LC_ADDRESS=C               LC_TELEPHONE=C            
#> [11] LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C       
#> 
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base     
#> 
#> other attached packages:
#> [1] memoise_1.1.0     modulr_0.1.7.9180
#> 
#> loaded via a namespace (and not attached):
#>  [1] igraph_1.0.1          Rcpp_0.12.11          rstudioapi_0.6       
#>  [4] whisker_0.3-2         knitr_1.16            xml2_1.1.1           
#>  [7] magrittr_1.5          roxygen2_6.0.1        devtools_1.12.0.9000 
#> [10] usethis_0.0.0.9000    pkgload_0.0.0.9000    R6_2.2.1             
#> [13] chorddiag_0.1.1       rlang_0.1.1.9000      pooh_0.3-2           
#> [16] httr_1.2.1            stringr_1.2.0         networkD3_0.2.13     
#> [19] tools_3.2.5           withr_1.0.2           htmltools_0.3.6      
#> [22] commonmark_1.2        yaml_2.1.14           assertthat_0.2.0.9000
#> [25] rprojroot_1.2         digest_0.6.12         tibble_1.3.3         
#> [28] pkgdown_0.1.0.9000    crayon_1.3.2          callr_1.0.0          
#> [31] purrr_0.2.2.2         htmlwidgets_0.8       testthat_1.0.2       
#> [34] evaluate_0.10         rmarkdown_1.5         stringi_1.1.5        
#> [37] desc_1.1.0            backports_1.1.0       highlight_0.4.7.1    
#> [40] jsonlite_1.5         

not_run({
  with_packages("~/my_packages", {
   if (!"pooh" %in% rownames(installed.packages()))
     utils::install.packages("pooh")
   library(pooh)
   print(sessionInfo())
  })
})
print(sessionInfo())

#> R version 3.2.5 (2016-04-14)
#> Platform: x86_64-pc-linux-gnu (64-bit)
#> Running under: Red Hat Enterprise Linux
#> 
#> locale:
#>  [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C              
#>  [3] LC_TIME=en_DK.UTF-8        LC_COLLATE=en_US.UTF-8    
#>  [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8   
#>  [7] LC_PAPER=en_US.UTF-8       LC_NAME=C                 
#>  [9] LC_ADDRESS=C               LC_TELEPHONE=C            
#> [11] LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C       
#> 
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base     
#> 
#> other attached packages:
#> [1] memoise_1.1.0     modulr_0.1.7.9180
#> 
#> loaded via a namespace (and not attached):
#>  [1] igraph_1.0.1          Rcpp_0.12.11          rstudioapi_0.6       
#>  [4] whisker_0.3-2         knitr_1.16            xml2_1.1.1           
#>  [7] magrittr_1.5          roxygen2_6.0.1        devtools_1.12.0.9000 
#> [10] usethis_0.0.0.9000    pkgload_0.0.0.9000    R6_2.2.1             
#> [13] chorddiag_0.1.1       rlang_0.1.1.9000      pooh_0.3-2           
#> [16] httr_1.2.1            stringr_1.2.0         networkD3_0.2.13     
#> [19] tools_3.2.5           withr_1.0.2           htmltools_0.3.6      
#> [22] commonmark_1.2        yaml_2.1.14           assertthat_0.2.0.9000
#> [25] rprojroot_1.2         digest_0.6.12         tibble_1.3.3         
#> [28] pkgdown_0.1.0.9000    crayon_1.3.2          callr_1.0.0          
#> [31] purrr_0.2.2.2         htmlwidgets_0.8       testthat_1.0.2       
#> [34] evaluate_0.10         rmarkdown_1.5         stringi_1.1.5        
#> [37] desc_1.1.0            backports_1.1.0       highlight_0.4.7.1    
#> [40] jsonlite_1.5         

## in file "foos/foobar.R"
"foos/foobar" %provides% {
 print(sessionInfo())
 with_module_packages({
   if (!"devtools" %in% rownames(installed.packages()))
     utils::install.packages("devtools")
   library(devtools)
   if (!"pooh" %in% rownames(installed.packages()))
     devtools::install_version("pooh", "0.2")
   library(pooh)
   print(sessionInfo())
 })
print(sessionInfo())
}

#> [2017-06-16T12:35:54 CEST] Defining 'foos/foobar' ... OK

## EOF
## Not run: make()

## in file "foos/foobaz.R"
"foos/foobaz" %provides% {
 print(sessionInfo())
 with_namespace_packages("foos", {
   if (!"devtools" %in% rownames(installed.packages()))
     utils::install.packages("devtools")
   library(devtools)
   if (!"pooh" %in% rownames(installed.packages()))
     devtools::install_version("pooh", "0.3")
   library(pooh)
   print(sessionInfo())
 })
print(sessionInfo())
}

#> [2017-06-16T12:35:54 CEST] Defining 'foos/foobaz' ... OK

## EOF
## Not run: make()